# openapi/bi-mvp.yaml
openapi: 3.1.0
info:
  title: Drag Visual BI MVP API
  version: 1.0.0
  description: Cross-language contract for dashboard drafts, publishing, and the dataset gateway.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:3000
security: []
...
paths:
  /dashboards:
    post:
      tags: [Dashboards]
      operationId: createDashboard
      summary: Create an empty dashboard draft
      requestBody:
        required: true
        x-maxUtf8Bytes: 2097152
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDashboardRequest' }
            example: { name: 销售看板 }
      responses:
        '201':
          description: Dashboard created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dashboard' }
              example: *dashboardExample
        '400': *dashboardSchemaInvalid
  /dashboards/{dashboardId}:
    parameters:
      - $ref: '#/components/parameters/DashboardId'
    get:
      tags: [Dashboards]
      operationId: getDashboard
      summary: Get a dashboard draft
      responses:
        '200':
          description: Current dashboard draft.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dashboard' }
              example: *dashboardExample
        '400': *dashboardSchemaInvalid
        '404': *dashboardNotFound
        '500': *internalError
    put:
      tags: [Dashboards]
      operationId: updateDashboard
      summary: Save a complete dashboard draft using optimistic revision
      requestBody:
        required: true
        x-maxUtf8Bytes: 2097152
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Dashboard' }
            example: *dashboardExample
      responses:
        '200':
          description: Saved dashboard with revision incremented exactly once.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dashboard' }
              example: *dashboardUpdatedExample
        '400': *dashboardSchemaInvalid
        '404': *dashboardNotFound
        '409': *dashboardConflict
        '500': *internalError
  /dashboards/{dashboardId}/publish:
    parameters:
      - $ref: '#/components/parameters/DashboardId'
    post:
      tags: [Dashboards]
      operationId: publishDashboard
      summary: Publish an immutable copy of the current draft
      description: Has no request body. Corrupt drafts and persistence failures retain the prior snapshot.
      responses:
        '200':
          description: Newly published dashboard snapshot.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dashboard' }
              example: *dashboardExample
        '400': *dashboardSchemaInvalid
        '404': *dashboardNotFound
        '500':
          description: Corrupt stored draft or transactional persistence failure; old snapshot is retained.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              examples:
                corruptDraft:
                  value: { code: INTERNAL_ERROR, message: Internal server error }
                persistenceFailure:
                  value: { code: PUBLISH_FAILED, message: Dashboard publish failed }
    delete:
      tags: [Dashboards]
      operationId: unpublishDashboard
      summary: Take a published dashboard page offline
      description: Removes the public snapshot while retaining the editable draft.
      responses:
        '200':
          description: The published page is no longer publicly available.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UnpublishResult' }
              example: { unpublished: true }
        '400': *dashboardSchemaInvalid
        '404': *dashboardNotFound
        '500': *internalError
  /published-dashboards/{dashboardId}:
    parameters:
      - $ref: '#/components/parameters/DashboardId'
    get:
      tags: [Dashboards]
      operationId: getPublishedDashboard
      summary: Get the last successful published snapshot
      responses:
        '200':
          description: Published dashboard snapshot.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dashboard' }
              example: *dashboardExample
        '400': *dashboardSchemaInvalid
        '404': *publishedNotFound
        '500': *internalError
  /datasets:
    get:
      tags: [Datasets]
      operationId: listDatasets
      summary: List allow-listed datasets without upstream details
      responses:
        '200':
          description: Dataset summaries without a wrapper.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DatasetSummary' }
              example: *datasetSummariesExample
        '500': *internalError
  /datasets/{datasetId}/schema:
    parameters:
      - $ref: '#/components/parameters/DatasetId'
    get:
      tags: [Datasets]
      operationId: getDatasetSchema
      summary: Get fields and query parameters for a dataset
      responses:
        '200':
          description: Dataset schema.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dataset' }
              example: *datasetExample
        '404': *datasetNotFound
        '502': *datasetBadGateway
  /datasets/{datasetId}/query:
    parameters:
      - $ref: '#/components/parameters/DatasetId'
    post:
      tags: [Datasets]
      operationId: queryDataset
      summary: Query an allow-listed dataset
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DatasetQueryRequest' }
            example: *queryRequestExample
      responses:
        '200':
          description: Normalized result, limited to 10,000 rows and 5 MiB UTF-8 JSON.
          x-maxUtf8Bytes: 5242880
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatasetQueryResult' }
              example: *queryResultExample
        '400': *datasetQueryInvalid
        '404': *datasetNotFound
        '502': *datasetBadGateway
        '504': *datasetTimeout
components:
  parameters:
    DashboardId:
      name: dashboardId
      in: path
      required: true
      description: Server-generated dashboard UUID.
      schema: { type: string, format: uuid }
      example: 6c614d7a-386b-4f36-a9ad-f9305b255b4f
    DatasetId:
      name: datasetId
      in: path
      required: true
      description: Stable opaque allow-list ID; it is not required to be a UUID.
      schema: { type: string, minLength: 1 }
      example: sales
  schemas:
    JsonValue:
      description: A recursive JSON-only value.
      anyOf:
        - type: 'null'
        - type: boolean
        - type: number
        - type: string
        - type: array
          items: { $ref: '#/components/schemas/JsonValue' }
        - type: object
          additionalProperties: { $ref: '#/components/schemas/JsonValue' }
    CreateDashboardRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: [string, 'null']
          maxLength: 100
          description: Missing, null, or trim-blank becomes 未命名看板.
    Theme:
      type: object
      additionalProperties: false
      required: [primaryColor, backgroundColor]
      properties:
        primaryColor: { type: string, pattern: '^#[0-9A-Fa-f]{6}$' }
        backgroundColor: { type: string, pattern: '^#[0-9A-Fa-f]{6}$' }
    GridItem:
      type: object
      additionalProperties: false
      required: [i, x, y, w, h]
      properties:
        i: { type: string, minLength: 1 }
        x: { type: integer, minimum: 0 }
        y: { type: integer, minimum: 0 }
        w: { type: integer, minimum: 1 }
        h: { type: integer, minimum: 1 }
    FieldBinding:
      type: object
      additionalProperties: false
      required: [fieldKey]
      properties:
        fieldKey: { type: string, minLength: 1 }
    SortBinding:
      type: object
      additionalProperties: false
      required: [fieldKey, direction]
      properties:
        fieldKey: { type: string, minLength: 1 }
        direction: { type: string, enum: [asc, desc] }
    DataBinding:
      type: object
      additionalProperties: false
      required: [datasetId, slots]
      properties:
        datasetId: { type: string, minLength: 1 }
        slots:
          type: object
          additionalProperties:
            oneOf:
              - $ref: '#/components/schemas/FieldBinding'
              - type: array
                items: { $ref: '#/components/schemas/FieldBinding' }
        sort: { $ref: '#/components/schemas/SortBinding' }
        limit: { type: integer, minimum: 1, maximum: 10000 }
    ComponentInstance:
      type: object
      additionalProperties: false
      required: [id, type, props]
      properties:
        id: { type: string, minLength: 1 }
        type: { type: string, enum: [bar, line, pie, kpi, table, text] }
        title: { type: string }
        props:
          type: object
          additionalProperties: { $ref: '#/components/schemas/JsonValue' }
        binding: { $ref: '#/components/schemas/DataBinding' }
    DashboardDataset:
      type: object
      additionalProperties: false
      required: [datasetId, schemaVersion, parameters]
      properties:
        datasetId: { type: string, minLength: 1 }
        schemaVersion: { type: string, minLength: 1 }
        parameters:
          type: object
          additionalProperties: { $ref: '#/components/schemas/JsonValue' }
    Dashboard:
      type: object
      additionalProperties: false
      x-maxUtf8Bytes: 2097152
      description: Component IDs, layout IDs, and dataset IDs are unique; each component has one matching layout item and bindings reference declared datasets.
      required: [schemaVersion, id, name, theme, layout, components, datasets, revision, updatedAt]
      properties:
        schemaVersion: { type: integer, const: 1 }
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 100 }
        theme: { $ref: '#/components/schemas/Theme' }
        layout:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/GridItem' }
        components:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/ComponentInstance' }
        datasets:
          type: array
          maxItems: 20
          items: { $ref: '#/components/schemas/DashboardDataset' }
        revision: { type: integer, minimum: 1 }
        updatedAt: { type: string, format: date-time }
        publishedAt: { type: [string, 'null'], format: date-time, readOnly: true }
      example: *dashboardExample
    UnpublishResult:
      type: object
      additionalProperties: false
      required: [unpublished]
      properties:
        unpublished: { type: boolean, const: true }
    DatasetField:
      type: object
      additionalProperties: false
      required: [key, label, type, nullable]
      properties:
        key: { type: string, minLength: 1 }
        label: { type: string, minLength: 1 }
        type: { type: string, enum: [string, number, date, boolean] }
        nullable: { type: boolean }
    QueryParameter:
      type: object
      additionalProperties: false
      required: [key, label, type, required]
      properties:
        key: { type: string, minLength: 1 }
        label: { type: string, minLength: 1 }
        type: { type: string, enum: [string, number, date, boolean] }
        required: { type: boolean }
    DatasetSummary:
      type: object
      additionalProperties: false
      required: [id, name, schemaVersion]
      properties:
        id: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        schemaVersion: { type: string, minLength: 1 }
    Dataset:
      type: object
      additionalProperties: false
      description: Field keys and query-parameter keys are unique within their respective arrays.
      required: [id, name, fields, parameters, schemaVersion]
      properties:
        id: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        fields:
          type: array
          items: { $ref: '#/components/schemas/DatasetField' }
        parameters:
          type: array
          items: { $ref: '#/components/schemas/QueryParameter' }
        schemaVersion: { type: string, minLength: 1 }
      example: *datasetExample
    DatasetQueryRequest:
      type: object
      additionalProperties: false
      description: Required parameters must be present and non-null; optional parameters may be omitted but cannot be null. Date values are calendar-valid YYYY-MM-DD strings and unknown parameters are rejected.
      required: [parameters]
      properties:
        parameters:
          type: object
          additionalProperties: { $ref: '#/components/schemas/JsonValue' }
      example: *queryRequestExample
    DatasetQueryResult:
      type: object
      additionalProperties: false
      x-maxUtf8Bytes: 5242880
      required: [columns, rows, sampledAt]
      properties:
        columns:
          type: array
          items: { $ref: '#/components/schemas/DatasetField' }
        rows:
          type: array
          maxItems: 10000
          items:
            type: object
            additionalProperties: { $ref: '#/components/schemas/JsonValue' }
        total: { type: integer, minimum: 0 }
        sampledAt: { type: string, format: date-time }
      description: Date values are calendar-valid YYYY-MM-DD strings; null is permitted only for nullable columns.
      example: *queryResultExample
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - DASHBOARD_SCHEMA_INVALID
            - DASHBOARD_NOT_FOUND
            - PUBLISHED_DASHBOARD_NOT_FOUND
            - DASHBOARD_ID_MISMATCH
            - DASHBOARD_VERSION_CONFLICT
            - DATASET_QUERY_INVALID
            - DATASET_NOT_FOUND
            - DATASET_INVALID_RESPONSE
            - DATASET_UPSTREAM_ERROR
            - DATASET_TIMEOUT
            - PUBLISH_FAILED
            - INTERNAL_ERROR
        message: { type: string, minLength: 1 }